Move search products API only used by Order into the OrderController
